requirements.txt
boto3==1.16.16
botocore==1.19.16
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
python-dateutil==2.8.1
python-dotenv==0.15.0
s3transfer==0.3.3
six==1.15.0
SQLAlchemy==1.3.20
urllib3==1.25.11
Werkzeug==1.0.1
WTForms==2.3.3
